requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9
PyJWT==2.8.0
flasgger==0.9.7.1
python-dotenv==1.0.1
